requirements.txt
ctransformers==0.2.5
sentence-transformers==2.6.1
pinecone-client==3.0.3
langchain==0.1.2
flask==3.0.2
python-dotenv
-e .
